mgr: Deal with long running avgs properly
We have recently started exposing long running avgs with the mgr python
modules but that only covers the total sum, not the avgcount.

With this patch, the long running avgs are exposed to the python code
as <counter>_sum and <counter>_count. This is also directly visible in
the ceph-mgr prometheus module.

Signed-off-by: Boris Ranto <branto@redhat.com>