prometheus: Fix metric resets
This patch changes the way we reset metrics when collecting data in the
prometheus exporter module. With this patch, the exported metrics at any
point in time should align with a freshly started ceph-mgr module.

The patch clears all the metrics at the beginning of scrape since these
can be easily outdated and were not cleared properly before. Afterwards
it directly sets the new labels based on the scraped data.

The patch also introduces a locking method when the process is trying to
scrape the metrics and a configurable caching mechanism.

This also removed the (now) outdated pending metrics logic.

Signed-off-by: Boris Ranto <branto@redhat.com>